+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/Annotator.java
@@ -37,16 +37,8 @@ import edu.ucr.cs.riple.core.evaluators.suppliers.Supplier;
 import edu.ucr.cs.riple.core.evaluators.suppliers.TargetModuleSupplier;
 import edu.ucr.cs.riple.core.injectors.AnnotationInjector;
 import edu.ucr.cs.riple.core.injectors.PhysicalInjector;
-import edu.ucr.cs.riple.core.metadata.field.FieldInitializationStore;
-import edu.ucr.cs.riple.core.metadata.index.Error;
 import edu.ucr.cs.riple.core.metadata.index.Fix;
 import edu.ucr.cs.riple.core.util.Utility;
-import edu.ucr.cs.riple.injector.changes.AddAnnotation;
-import edu.ucr.cs.riple.injector.changes.AddMarkerAnnotation;
-import edu.ucr.cs.riple.injector.changes.AddSingleElementAnnotation;
-import edu.ucr.cs.riple.injector.location.OnField;
-import edu.ucr.cs.riple.injector.location.OnParameter;
-import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -94,17 +86,7 @@ public class Annotator {
    */
   private void preprocess() {
     System.out.println("Preprocessing...");
-    Set<OnField> uninitializedFields =
-        Utility.readFixesFromOutputDirectory(context, context.targetModuleInfo).stream()
-            .filter(fix -> fix.isOnField() && fix.reasons.contains("FIELD_NO_INIT"))
-            .map(Fix::toField)
-            .collect(Collectors.toSet());
-    FieldInitializationStore fieldInitializationStore = new FieldInitializationStore(context);
-    Set<AddAnnotation> initializers =
-        fieldInitializationStore.findInitializers(uninitializedFields).stream()
-            .map(onMethod -> new AddMarkerAnnotation(onMethod, config.initializerAnnot))
-            .collect(Collectors.toSet());
-    this.injector.injectAnnotations(initializers);
+    context.checker.preprocess(injector);
   }
 
   /** Performs iterations of inference/injection until no unseen fix is suggested. */
@@ -137,8 +119,8 @@ public class Annotator {
         cache.enable();
       }
     }
-    if (config.forceResolveActivated) {
-      forceResolveRemainingErrors();
+    if (config.suppressRemainingErrors) {
+      context.checker.suppressRemainingErrors(injector);
     }
     System.out.println("\nFinished annotating.");
     Utility.writeReports(context, cache.reports().stream().collect(ImmutableSet.toImmutableSet()));
@@ -218,150 +200,5 @@ public class Annotator {
       return new CachedEvaluator(supplier);
     }
     return new BasicEvaluator(supplier);
-  }
-
-  /**
-   * Resolves all remaining errors in target module by following steps below:
-   *
-   * <ul>
-   *   <li>Enclosing method of triggered errors will be marked with {@code @NullUnmarked}
-   *       annotation.
-   *   <li>Uninitialized fields (inline or by constructor) will be annotated as
-   *       {@code @SuppressWarnings("NullAway.Init")}.
-   *   <li>Explicit {@code Nullable} assignments to fields will be annotated as
-   *       {@code @SuppressWarnings("NullAway")}.
-   * </ul>
-   */
-  private void forceResolveRemainingErrors() {
-    // Collect regions with remaining errors.
-    Utility.buildTarget(context);
-    Set<Error> remainingErrors =
-        Utility.readErrorsFromOutputDirectory(context, context.targetModuleInfo);
-    Set<Fix> remainingFixes =
-        Utility.readFixesFromOutputDirectory(context, context.targetModuleInfo);
-    // Collect all regions for NullUnmarked.
-    // For all errors in regions which correspond to a method's body, we can add @NullUnmarked at
-    // the method level.
-    Set<AddAnnotation> nullUnMarkedAnnotations =
-        remainingErrors.stream()
-            // find the corresponding method nodes.
-            .map(
-                error -> {
-                  if (error.getRegion().isOnCallable()
-                      &&
-                      // We suppress initialization errors reported on constructors using
-                      // @SuppressWarnings("NullAway.Init"). We add @NullUnmarked on constructors
-                      // only for errors in the body of the constructor.
-                      !error.isInitializationError()) {
-                    return context
-                        .targetModuleInfo
-                        .getMethodRegistry()
-                        .findMethodByName(error.encClass(), error.encMember());
-                  }
-                  // For methods invoked in an initialization region, where the error is that
-                  // `@Nullable` is being passed as an argument, we add a `@NullUnmarked` annotation
-                  // to the called method.
-                  if (error.messageType.equals("PASS_NULLABLE")
-                      && error.isSingleFix()
-                      && error.toResolvingLocation().isOnParameter()) {
-                    OnParameter nullableParameter = error.toResolvingParameter();
-                    return context
-                        .targetModuleInfo
-                        .getMethodRegistry()
-                        .findMethodByName(
-                            nullableParameter.clazz, nullableParameter.enclosingMethod.method);
-                  }
-                  return null;
-                })
-            // Filter null values from map above.
-            .filter(Objects::nonNull)
-            .map(node -> new AddMarkerAnnotation(node.location, config.nullUnMarkedAnnotation))
-            .collect(Collectors.toSet());
-
-    // For errors within static initialization blocks, add a @NullUnmarked annotation on the
-    // enclosing class
-    nullUnMarkedAnnotations.addAll(
-        remainingErrors.stream()
-            .filter(
-                error ->
-                    error.getRegion().isOnInitializationBlock()
-                        && !error.getRegion().isInAnonymousClass())
-            .map(
-                error ->
-                    new AddMarkerAnnotation(
-                        context.targetModuleInfo.getLocationOnClass(error.getRegion().clazz),
-                        config.nullUnMarkedAnnotation))
-            .collect(Collectors.toSet()));
-    injector.injectAnnotations(nullUnMarkedAnnotations);
-    // Update log.
-    context.log.updateInjectedAnnotations(nullUnMarkedAnnotations);
-
-    // Collect suppress warnings, errors on field declaration regions.
-    Set<OnField> fieldsWithSuppressWarnings =
-        remainingErrors.stream()
-            .filter(
-                error -> {
-                  if (!error.getRegion().isOnField()) {
-                    return false;
-                  }
-                  if (error.messageType.equals("PASS_NULLABLE")) {
-                    // It is already resolved with @NullUnmarked selected above.
-                    return false;
-                  }
-                  // We can silence them by SuppressWarnings("NullAway.Init")
-                  return !error.isInitializationError();
-                })
-            .map(
-                error ->
-                    context
-                        .targetModuleInfo
-                        .getFieldRegistry()
-                        .getLocationOnField(error.getRegion().clazz, error.getRegion().member))
-            .filter(Objects::nonNull)
-            .collect(Collectors.toSet());
-
-    Set<AddAnnotation> suppressWarningsAnnotations =
-        fieldsWithSuppressWarnings.stream()
-            .map(
-                onField ->
-                    new AddSingleElementAnnotation(onField, "SuppressWarnings", "NullAway", false))
-            .collect(Collectors.toSet());
-    injector.injectAnnotations(suppressWarningsAnnotations);
-    // Update log.
-    context.log.updateInjectedAnnotations(suppressWarningsAnnotations);
-
-    // Collect NullAway.Init SuppressWarnings
-    Set<AddAnnotation> initializationSuppressWarningsAnnotations =
-        remainingFixes.stream()
-            .filter(
-                fix ->
-                    fix.isOnField()
-                        && (fix.reasons.contains("METHOD_NO_INIT")
-                            || fix.reasons.contains("FIELD_NO_INIT")))
-            // Filter nodes annotated with SuppressWarnings("NullAway")
-            .filter(fix -> !fieldsWithSuppressWarnings.contains(fix.toField()))
-            .map(
-                fix ->
-                    new AddSingleElementAnnotation(
-                        fix.toField(), "SuppressWarnings", "NullAway.Init", false))
-            .collect(Collectors.toSet());
-    injector.injectAnnotations(initializationSuppressWarningsAnnotations);
-    // Update log.
-    context.log.updateInjectedAnnotations(initializationSuppressWarningsAnnotations);
-    // Collect @NullUnmarked annotations on classes for any remaining error.
-    Utility.buildTarget(context);
-    remainingErrors = Utility.readErrorsFromOutputDirectory(context, context.targetModuleInfo);
-    nullUnMarkedAnnotations =
-        remainingErrors.stream()
-            .filter(error -> !error.getRegion().isInAnonymousClass())
-            .map(
-                error ->
-                    new AddMarkerAnnotation(
-                        context.targetModuleInfo.getLocationOnClass(error.getRegion().clazz),
-                        config.nullUnMarkedAnnotation))
-            .collect(Collectors.toSet());
-    injector.injectAnnotations(nullUnMarkedAnnotations);
-    // Update log.
-    context.log.updateInjectedAnnotations(nullUnMarkedAnnotations);
   }
 }
