+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/Config.java
@@ -27,7 +27,6 @@ package edu.ucr.cs.riple.core;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Sets;
-import edu.ucr.cs.riple.core.io.deserializers.CheckerDeserializer;
 import edu.ucr.cs.riple.core.module.ModuleConfiguration;
 import edu.ucr.cs.riple.core.util.Utility;
 import edu.ucr.cs.riple.scanner.generatedcode.SourceType;
@@ -73,8 +72,8 @@ public class Config {
   /** If activated, impact of fixes will be cached. */
   public final boolean useImpactCache;
   /**
-   * If activated, all suggested fixes from NullAway will be applied to the source code regardless
-   * of their effectiveness.
+   * If activated, all suggested fixes from the checker will be applied to the source code
+   * regardless of their effectiveness.
    */
   public final boolean exhaustiveSearch;
   /**
@@ -93,7 +92,7 @@ public class Config {
   public final boolean disableOuterLoop;
   /** Info of target module. */
   public final ModuleConfiguration target;
-  /** Path to directory where all outputs of nullaway and scanner checker is located. */
+  /** Path to directory where all outputs of checker and scanner checker is located. */
   public final Path globalDir;
   /** Command to build the target module. */
   public final String buildCommand;
@@ -106,8 +105,8 @@ public class Config {
   /** Sets of context path information for all downstream dependencies. */
   public final ImmutableSet<ModuleConfiguration> downstreamConfigurations;
   /**
-   * Path to nullaway library model loader, which enables the communication between annotator and
-   * nullaway when processing downstream dependencies.
+   * Path to NullAway library model loader, which enables the communication between annotator and
+   * NullAway when processing downstream dependencies.
    */
   public final Path nullawayLibraryModelLoaderPath;
   /** Command to build the all downstream dependencies at once. */
@@ -120,11 +119,11 @@ public class Config {
   /** Global counter for assigning unique id for each instance. */
   private int moduleCounterID;
   /**
-   * If activated, AutoAnnotator will try to resolve all remaining errors by marking the enclosing
-   * method as {@code NullUnMarked}. It will also mark uninitialized fields with
-   * {@code @SuppressWarning("NullAway.Init")}
+   * If activated, Annotator will try to suppress all remaining errors. The logic is specific to the
+   * checker suppression mechanism. Annotator will simply call {@link
+   * edu.ucr.cs.riple.core.checkers.Checker#suppressRemainingErrors} methods.
    */
-  public final boolean forceResolveActivated;
+  public final boolean suppressRemainingErrors;
   /** Fully qualified NullUnmarked annotation. */
   public final String nullUnMarkedAnnotation;
   /**
@@ -147,8 +146,10 @@ public class Config {
    * code and are generated by a processor.
    */
   public final ImmutableSet<SourceType> generatedCodeDetectors;
-  /** Checker enum to retrieve checker specific instances. (e.g. {@link CheckerDeserializer}) */
-  public final Checker checker;
+  /**
+   * Checker name to retrieve the {@link edu.ucr.cs.riple.core.checkers.Checker} specific instance.
+   */
+  public final String checkerName;
 
   /**
    * Builds context from command line arguments.
@@ -185,9 +186,9 @@ public class Config {
     Option configPathsOption =
         new Option(
             "cp",
-            "context-paths",
+            "config-paths",
             true,
-            "Path to tsv file containing path to nullaway and scanner context files.");
+            "Path to tsv file containing path to checker and scanner context files.");
     configPathsOption.setRequired(true);
     options.addOption(configPathsOption);
 
@@ -381,7 +382,7 @@ public class Config {
         cmd.hasOption(nullableOption.getLongOpt())
             ? cmd.getOptionValue(nullableOption.getLongOpt())
             : "javax.annotation.Nullable";
-    this.checker = Checker.getCheckerByName(cmd.getOptionValue(checkerNameOption));
+    this.checkerName = cmd.getOptionValue(checkerNameOption);
     this.initializerAnnot = cmd.getOptionValue(initializerOption.getLongOpt());
     this.depth =
         Integer.parseInt(
@@ -435,10 +436,10 @@ public class Config {
       this.downstreamDependenciesBuildCommand = null;
     }
     this.inferenceActivated = !cmd.hasOption(deactivateInference);
-    this.forceResolveActivated =
+    this.suppressRemainingErrors =
         !this.inferenceActivated || cmd.hasOption(activateForceResolveOption);
     this.nullUnMarkedAnnotation =
-        this.forceResolveActivated
+        this.suppressRemainingErrors
             ? cmd.getOptionValue(activateForceResolveOption)
             : "org.jspecify.annotations.NullUnmarked";
     this.moduleCounterID = 0;
@@ -467,8 +468,7 @@ public class Config {
     } catch (Exception e) {
       throw new RuntimeException("Error in reading/parsing context at path: " + configPath, e);
     }
-    this.checker =
-        Checker.getCheckerByName(getValueFromKey(jsonObject, "CHECKER", String.class).orElse(null));
+    this.checkerName = getValueFromKey(jsonObject, "CHECKER", String.class).orElse(null);
     this.depth = getValueFromKey(jsonObject, "DEPTH", Long.class).orElse((long) 1).intValue();
     this.chain = getValueFromKey(jsonObject, "CHAIN", Boolean.class).orElse(false);
     this.redirectBuildOutputToStdErr =
@@ -526,7 +526,7 @@ public class Config {
 
     this.downstreamConfigurations = ImmutableSet.copyOf(moduleConfigurationList);
     this.moduleCounterID = 0;
-    this.forceResolveActivated =
+    this.suppressRemainingErrors =
         getValueFromKey(jsonObject, "FORCE_RESOLVE", Boolean.class).orElse(false);
     this.inferenceActivated =
         getValueFromKey(jsonObject, "INFERENCE_ACTIVATION", Boolean.class).orElse(true);
@@ -650,7 +650,6 @@ public class Config {
     public String initializerAnnotation;
     public String nullableAnnotation;
     public String outputDir;
-    public Checker checker;
     /**
      * List of modules, did not use {@link java.util.Set} to preserve order. First project is the
      * target project.
@@ -674,6 +673,7 @@ public class Config {
     public boolean useCacheImpact = false;
     public Set<SourceType> sourceTypes = new HashSet<>();
     public int depth = 1;
+    public String checker;
 
     @SuppressWarnings("unchecked")
     public void write(Path path) {
@@ -688,7 +688,7 @@ public class Config {
           nullableAnnotation, "Nullable Annotation must be initialized to construct the context.");
       JSONObject json = new JSONObject();
       json.put("BUILD_COMMAND", buildCommand);
-      json.put("CHECKER", checker.name());
+      json.put("CHECKER", checker);
       JSONObject annotation = new JSONObject();
       annotation.put("INITIALIZER", initializerAnnotation);
       annotation.put("NULLABLE", nullableAnnotation);
@@ -712,7 +712,7 @@ public class Config {
               .map(
                   info -> {
                     JSONObject res = new JSONObject();
-                    res.put("NULLAWAY", info.nullawayConfig.toString());
+                    res.put("CHECKER", info.checkerConfig.toString());
                     res.put("SCANNER", info.scannerConfig.toString());
                     return res;
                   })
