+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/util/Utility.java
@@ -40,7 +40,6 @@ import edu.ucr.cs.riple.scanner.ScannerConfigWriter;
 import edu.ucr.cs.riple.scanner.generatedcode.SourceType;
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
-import java.io.File;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.nio.charset.Charset;
@@ -51,23 +50,12 @@ import java.time.temporal.ChronoUnit;
 import java.util.Collections;
 import java.util.List;
 import java.util.Set;
-import java.util.UUID;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import javax.xml.parsers.DocumentBuilder;
-import javax.xml.parsers.DocumentBuilderFactory;
-import javax.xml.parsers.ParserConfigurationException;
-import javax.xml.transform.Transformer;
-import javax.xml.transform.TransformerException;
-import javax.xml.transform.TransformerFactory;
-import javax.xml.transform.dom.DOMSource;
-import javax.xml.transform.stream.StreamResult;
 import me.tongfei.progressbar.ProgressBar;
 import me.tongfei.progressbar.ProgressBarStyle;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
-import org.w3c.dom.Document;
-import org.w3c.dom.Element;
 
 /** Utility class. */
 public class Utility {
@@ -143,13 +131,12 @@ public class Utility {
    * Reads serialized errors "errors.tsv" file in the output directory, and returns the collected
    * set of resolving fixes for read errors.
    *
-   * @param context Annotator context. Required to fetch the deserializer.
-   * @param moduleInfo ModuleInfo of the module which fixes are created for.
+   * @param context Context of the annotator.
+   * @param moduleInfo ModuleInfo of the module which errors are created for.
    * @return Set of collected fixes.
    */
   public static Set<Fix> readFixesFromOutputDirectory(Context context, ModuleInfo moduleInfo) {
-    Set<Error> errors = readErrorsFromOutputDirectory(context, moduleInfo);
-    return Error.getResolvingFixesOfErrors(errors);
+    return Error.getResolvingFixesOfErrors(context.checker.deserializeErrors(moduleInfo));
   }
 
   /**
@@ -159,73 +146,11 @@ public class Utility {
    * @param moduleInfo ModuleInfo of the module which errors are created for.
    * @return Set of serialized errors.
    */
-  public static Set<Error> readErrorsFromOutputDirectory(Context context, ModuleInfo moduleInfo) {
-    return context.deserializer.deserializeErrors(moduleInfo);
-  }
-
-  /**
-   * Writes the {@link FixSerializationConfig} in {@code XML} format.
-   *
-   * @param config Context file to write.
-   * @param path Path to write the context at.
-   */
-  public static void writeNullAwayConfigInXMLFormat(FixSerializationConfig config, String path) {
-    DocumentBuilderFactory docFactory = DocumentBuilderFactory.newInstance();
-    try {
-      DocumentBuilder docBuilder = docFactory.newDocumentBuilder();
-      Document doc = docBuilder.newDocument();
-
-      // Root
-      Element rootElement = doc.createElement("serialization");
-      doc.appendChild(rootElement);
-
-      // Suggest
-      Element suggestElement = doc.createElement("suggest");
-      suggestElement.setAttribute("active", String.valueOf(config.suggestEnabled));
-      suggestElement.setAttribute("enclosing", String.valueOf(config.suggestEnclosing));
-      rootElement.appendChild(suggestElement);
-
-      // Field Initialization
-      Element fieldInitInfoEnabled = doc.createElement("fieldInitInfo");
-      fieldInitInfoEnabled.setAttribute("active", String.valueOf(config.fieldInitInfoEnabled));
-      rootElement.appendChild(fieldInitInfoEnabled);
-
-      // Output dir
-      Element outputDir = doc.createElement("path");
-      outputDir.setTextContent(config.outputDirectory);
-      rootElement.appendChild(outputDir);
-
-      // UUID
-      Element uuid = doc.createElement("uuid");
-      uuid.setTextContent(UUID.randomUUID().toString());
-      rootElement.appendChild(uuid);
-
-      // Writings
-      TransformerFactory transformerFactory = TransformerFactory.newInstance();
-      Transformer transformer = transformerFactory.newTransformer();
-      DOMSource source = new DOMSource(doc);
-      StreamResult result = new StreamResult(new File(path));
-      transformer.transform(source, result);
-    } catch (ParserConfigurationException | TransformerException e) {
-      throw new RuntimeException("Error happened in writing config.", e);
-    }
-  }
-
-  /**
-   * Enables NullAway serialization for the given modules.
-   *
-   * @param configurations Set of modules to enable NullAway serialization for.
-   */
-  public static void enableNullAwaySerialization(ImmutableSet<ModuleConfiguration> configurations) {
-    configurations.forEach(
-        module -> {
-          FixSerializationConfig.Builder nullAwayConfig =
-              new FixSerializationConfig.Builder()
-                  .setSuggest(true, true)
-                  .setOutputDirectory(module.dir.toString())
-                  .setFieldInitInfo(true);
-          nullAwayConfig.writeAsXML(module.nullawayConfig.toString());
-        });
+  public static <T extends Error> Set<T> readErrorsFromOutputDirectory(
+      Context context, ModuleInfo moduleInfo, Class<T> klass) {
+    return context.checker.deserializeErrors(moduleInfo).stream()
+        .map(klass::cast)
+        .collect(Collectors.toSet());
   }
 
   /**
@@ -296,40 +221,17 @@ public class Utility {
    * @param context Annotator context.
    */
   public static void buildDownstreamDependencies(Context context) {
-    context.downstreamConfigurations.forEach(
-        module -> {
-          FixSerializationConfig.Builder nullAwayConfig =
-              new FixSerializationConfig.Builder()
-                  .setSuggest(true, true)
-                  .setOutputDirectory(module.dir.toString())
-                  .setFieldInitInfo(false);
-          nullAwayConfig.writeAsXML(module.nullawayConfig.toString());
-        });
+    context.checker.prepareConfigFilesForBuild(context.downstreamConfigurations);
     build(context, context.config.downstreamDependenciesBuildCommand);
-  }
-
-  /**
-   * Builds target.
-   *
-   * @param context Annotator context.
-   */
-  public static void buildTarget(Context context) {
-    buildTarget(context, false);
   }
 
   /**
    * Builds target with control on field initialization serialization.
    *
    * @param context Annotator context.
-   * @param initSerializationEnabled Activation flag for field initialization serialization.
    */
-  public static void buildTarget(Context context, boolean initSerializationEnabled) {
-    FixSerializationConfig.Builder nullAwayConfig =
-        new FixSerializationConfig.Builder()
-            .setSuggest(true, true)
-            .setOutputDirectory(context.targetConfiguration.dir.toString())
-            .setFieldInitInfo(initSerializationEnabled);
-    nullAwayConfig.writeAsXML(context.targetConfiguration.nullawayConfig.toString());
+  public static void buildTarget(Context context) {
+    context.checker.prepareConfigFilesForBuild(context.targetModuleInfo.getModuleConfigurations());
     build(context, context.config.buildCommand);
   }
